audiomp3: rename to MP3Decoder
