+++ shared-module/audiomp3/MP3Decoder.c
